requirements.txt
langchain==0.2.6
langchain-community==0.2.6
pypdfium2==4.30.0
streamlit==1.36.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1